add LCanvas component, migrate canvas interaction and rendering business logic from graphcomponent to new module
