+++ src/components/GraphComponent/GraphComponent.js
@@ -1,13 +1,13 @@
 import React, { Component } from 'react';
 import './GraphComponent.css';
 import InputField from '../InputField/InputField';
+import LCanvas from '../LCanvas/LCanvas';
 import Turtle from '../../services/turtle';
 import config from '../../config/config.json';
 
 const {
   RESERVED_CHARS,
   MAX_ITERATIONS,
-  SCALE_FACTOR,
 } = config;
 
 /**
@@ -19,11 +19,6 @@ class GraphComponent extends Component {
     this.state = {
       clientWidth: 0,
       clientHeight: 0,
-      offsetX: 0,
-      offsetY: 0,
-      fOffsetX: 0,
-      fOffsetY: 0,
-      context: null,
       initPath: 'FX',
       replaceFn: [
         {
@@ -58,20 +53,13 @@ class GraphComponent extends Component {
       stepLength: 50,
       alpha: 90,
       iteration: 0,
-      scale: 1,
     };
 
-    this.animate = this.animate.bind(this);
-    this.updateLines = this.updateLines.bind(this);
     this.updateReplaceFns = this.updateReplaceFns.bind(this);
     this.handleChange = this.handleChange.bind(this);
-    this.handleDragStart = this.handleDragStart.bind(this);
-    this.handleDrag = this.handleDrag.bind(this);
-    this.handleDragEnd = this.handleDragEnd.bind(this);
     this.handleReplaceFn = this.handleReplaceFn.bind(this);
     this.handleReplaceFnToggle = this.handleReplaceFnToggle.bind(this);
     this.handleResize = this.handleResize.bind(this);
-    this.handleScroll = this.handleScroll.bind(this);
   }
 
   /**
@@ -80,17 +68,12 @@ class GraphComponent extends Component {
    */
   componentDidMount() {
     window.addEventListener('resize', this.handleResize);
-    window.addEventListener('wheel', this.handleScroll);
     const { clientWidth, clientHeight } = this.wrapper;
     const turtle = new Turtle(clientWidth, clientHeight);
-    const context = this.canvas.getContext('2d');
     this.setState({
-      context,
       clientWidth,
       clientHeight,
       turtle,
-      fOffsetX: 0,
-      fOffsetY: 0,
     }, this.updateReplaceFns);
   }
 
@@ -103,64 +86,17 @@ class GraphComponent extends Component {
   }
 
   /**
-   * Clears canvas and draws lines onto it.
+   * Updates replacement functions and line objects.
    */
-  animate() {
+  updateReplaceFns() {
     const {
-      lines,
-      context,
-      clientWidth,
-      clientHeight,
-      offsetX,
-      offsetY,
-      fOffsetX,
-      fOffsetY,
-      scale,
-    } = this.state;
-    context.clearRect(0, 0, clientWidth, clientHeight);
-    lines.forEach((line) => {
-      context.beginPath();
-      context.moveTo(
-        scale * (offsetX + fOffsetX + line.start.x),
-        scale * (offsetY + fOffsetY + line.start.y),
-      );
-      context.lineTo(
-        scale * (offsetX + fOffsetX + line.end.x),
-        scale * (offsetY + fOffsetY + line.end.y),
-      );
-      context.stroke();
-    });
-  }
-
-  /**
-   * Parses current state and updates line objects.
-   */
-  updateLines() {
-    const {
-      turtle,
-      replaceFn,
       initPath,
+      replaceFn,
+      turtle,
       iteration,
       alpha,
       stepLength,
     } = this.state;
-    this.setState({
-      lines: turtle.parse(
-        initPath,
-        replaceFn,
-        iteration,
-        Number(alpha),
-        stepLength,
-      ),
-    }, this.animate);
-  }
-
-  /**
-   * Adds replacement string entries if a new character has been typed. Removes
-   * replacement strings whose chars no longer have any occurrences.
-   */
-  updateReplaceFns() {
-    const { initPath, replaceFn } = this.state;
     const chars = replaceFn.map(rF => rF.char);
     const initChars = initPath.split('');
     replaceFn.forEach((rF) => {
@@ -189,7 +125,16 @@ class GraphComponent extends Component {
         replaceFn.splice(j, 1);
       }
     });
-    this.setState({ replaceFn }, this.updateLines);
+
+    this.setState({
+      lines: turtle.parse(
+        initPath,
+        replaceFn,
+        iteration,
+        Number(alpha),
+        stepLength,
+      ),
+    });
   }
 
   /**
@@ -224,51 +169,6 @@ class GraphComponent extends Component {
   }
 
   /**
-   * Storese initial drag location and adds event listener for mouse movement.
-   * @param {Event} e - User input event.
-   */
-  handleDragStart(e) {
-    const position = this.canvas.getBoundingClientRect();
-    this.setState({
-      startX: e.pageX - position.x,
-      startY: e.pageY - position.y,
-    });
-    window.addEventListener('mousemove', this.handleDrag);
-  }
-
-  /**
-   * Updates drag offset distance.
-   * @param {Event} e - User input event.
-   */
-  handleDrag(e) {
-    const { scale, startX, startY } = this.state;
-    const position = this.canvas.getBoundingClientRect();
-    const offsetX = (e.pageX - position.x - startX) / scale;
-    const offsetY = (e.pageY - position.y - startY) / scale;
-    this.setState({ offsetX, offsetY }, this.animate);
-  }
-
-  /**
-   * Copies current drag offset distance to fixed offset distance, then sets
-   * current drag offset distance to (0,0).
-   */
-  handleDragEnd() {
-    const {
-      offsetX,
-      offsetY,
-      fOffsetX,
-      fOffsetY,
-    } = this.state;
-    this.setState({
-      fOffsetX: offsetX + fOffsetX,
-      fOffsetY: offsetY + fOffsetY,
-      offsetX: 0,
-      offsetY: 0,
-    });
-    window.removeEventListener('mousemove', this.handleDrag);
-  }
-
-  /**
    * Updates canvas dimensions to match new window dimensions and initializes a
    * new turtle.
    */
@@ -279,20 +179,7 @@ class GraphComponent extends Component {
       clientWidth,
       clientHeight,
       turtle,
-    }, this.animate);
-  }
-
-  /**
-   * Updates the scale factor on user wheel.
-   * @param {Event} e - User input event.
-   */
-  handleScroll(e) {
-    const { scale } = this.state;
-    this.setState({
-      scale: e.deltaY < 0
-        ? scale * SCALE_FACTOR
-        : scale / SCALE_FACTOR,
-    }, this.animate);
+    });
   }
 
   render() {
@@ -304,21 +191,18 @@ class GraphComponent extends Component {
       stepLength,
       clientWidth,
       clientHeight,
+      lines,
     } = this.state;
 
     return (
       <div className="graph-wrapper">
         <div className="graph-panel left">
           <div
-            tabIndex={0}
-            role="button"
-            className="graph-canvas-wrapper"
             ref={(node) => { this.wrapper = node; }}
-            onMouseDown={this.handleDragStart}
-            onMouseUp={this.handleDragEnd}
+            className="graph-canvas-wrapper"
           >
-            <canvas
-              ref={(node) => { this.canvas = node; }}
+            <LCanvas
+              lines={lines}
               width={clientWidth}
               height={clientHeight}
             />
