+++ src/components/LCanvas/LCanvas.scss
@@ -1,0 +1,17 @@
+.L-canvas-wrapper {
+    width: 100%;
+    height: 100%;
+    overflow: visible;
+    padding: 4px;
+}
+
+.L-canvas {
+    background-color: #{$theme-white};
+    width: calc(100% - 8px);
+    border-radius: 2px;
+    box-sizing: border-box;
+    box-shadow: 
+        1px 1px rgba(0,0,0,0.1),
+        1.5px 1.5px rgba(0,0,0,0.17),
+        2px 2px rgba(0,0,0,0.24);
+}
